Run gofmt on all generated Go sources
Currently the code generator handles expliclty the begin and end of
blocks, in order to generate nicely indented code. But Go has the gofmt
tool, which can automatically re-arrange the code. This patch changes
the code generator so that it doesn't handle blocks explicitly, instead
it runs the gofmt tool on the generated code.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>