Don't sort types topologically
In the Python and Ruby SDKs the declarations of the types need to be
sorted topologically, because the language requires declarations of base
types before derived types. But in Go this isn't necessary. This patch
takes advantage of that to simplify a bit the generator of the code for
the types.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>