Add .travis.yml
This patch modifies the POM file of the SDK so that during the compile
phase it compiles all the Go source code. Then it adds a .travis.yml
file so that this compilation will be also done in Travis CI.

Note that the .travis.yml file is currently configured to skip the
tests, because one of the tests tries to use HTTP communication to an
oVirt engine server, which will always fail in the Travis CI
environment.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>