Better error handling for HTTP 500 errors
This change adds better handling for HTTP 500 / non-XML errors (e.g. HTML error messages). It refactors the HTTP body reading and provides the response as a slice instead of having it read in several places, possibly resulting in multiple attempts to read the body. The resulting error is given a type (`ResponseParseError`), which can be used to identify errors when the response is not an XML response.