Fix Apache 2.0 license text
The Go package documentation website at go.pkg.dev only contains
packages that contain a known OSI approved license. The document
included in the oVirt SDK is missing the appendix, causing a large
enough difference percentage to mismatch.

The LICENSE file in this repository already contains the correct license
text (including the appendix), the LICENSE.txt file that is copied to
the SDK does not.

https://pkg.go.dev/github.com/ovirt/go-ovirt

https://pkg.go.dev/license-policy

https://pkg.go.dev/github.com/google/licensecheck

Signed-off-by: Vic Demuzere <vic@demuzere.be>